Added aspects, logging, stock scheduling, enhanced excption handling, more features

+++ src/main/java/com/example/store/entities/CartItem.java
@@ -33,7 +33,7 @@ public class CartItem {
     @Column(name = "quantity")
     private Long quantity;
 
-    public Long getTotalPrice() {
+    public Double getTotalPrice() {
         return product.getPrice() * quantity;
     }
